+++ src/widget/mod.rs
@@ -1,4 +1,5 @@
 pub mod color_editors;
+pub mod combo_box;
 pub mod image;
 pub mod misc;
 pub mod progress_bar;